requirements.txt
sanic==21.9.3
PyJWT==2.3.0